Add save category feature

+++ inventory-service/src/main/java/com/project/inventory/dto/ProductRecord.java
@@ -3,6 +3,8 @@ package com.project.inventory.dto;
 import com.project.inventory.entity.Category;
 import com.project.inventory.entity.SubCategory;
 
-public record ProductRecord(Long id, String name, String description, double price, int quantity, String imageUrl,
+import java.math.BigDecimal;
+
+public record ProductRecord(Long id, String name, String description, BigDecimal price, int quantity, String imageUrl,
                             Category category, SubCategory subCategory) {
 }
